// packages/core/src/types.ts
export interface TypedDocumentNode<
  Result = { [key: string]: any },
  Variables = { [key: string]: any }
> extends DocumentNode {
  /** Type to check whether `Variables` and `Result` are assignable types.
   * @internal
   */
  __apiType?: (variables: Variables) => Result;
}
// ...
export type ExecutionResult =
  | {
      /** Contains a list of errors raised by fields or the request itself.
       * @see {@link https://spec.graphql.org/October2021/#sec-Errors} for the GraphQL Errors Response spec
       */
      errors?:
        | Array<Partial<GraphQLError> | string | Error>
        | readonly GraphQLError[];
      /** The result of the execution of the GraphQL operation.
       * @see {@link https://spec.graphql.org/October2021/#sec-Data} for the GraphQL Data Response spec
       */
      data?: null | Record<string, any>;
      /** Additional metadata that a GraphQL API may choose to send that is out of spec.
       * @see {@link https://spec.graphql.org/October2021/#sel-EAPHJCAACCoGu9J} for the GraphQL Response spec
       */
      extensions?: Record<string, any>;
      /** Flag indicating whether a future, incremental response may update this response.
       * @see {@link https://github.com/graphql/graphql-wg/blob/main/rfcs/DeferStream.md#payload-format} for the DeferStream spec
       */
      hasNext?: boolean;
    }
  | {
      /** Contains a list of errors raised by fields or the request itself.
       * @see {@link https://spec.graphql.org/October2021/#sec-Errors} for the GraphQL Errors Response specification
       */
      errors?:
        | Array<Partial<GraphQLError> | string | Error>
        | readonly GraphQLError[];
      /** JSON data that replaces existing data as part of an incremental response.
       * @see {@link https://github.com/graphql/graphql-wg/blob/main/rfcs/DeferStream.md#payload-format} for the DeferStream spec
       */
      data: any;
      /** JSON path at which the `data` should be inserted or replace existing data.
       *
       * @remarks
       * The `data` of an incremental response is inserted at `path` to update a previous result, as indicated when `hasNext: true`
       * was set previously. This means that an incremental GraphQL response can be updated with future response payloads, changing
       * or completing a GraphQL operation over time. This is used for `@defer`, `@stream`, and `@live` queries.
       *
       * @see {@link https://github.com/graphql/graphql-wg/blob/main/rfcs/DeferStream.md#payload-format} for the DeferStream spec
       */
      path: (string | number)[];
      /** Flag indicating whether a future, incremental response may update this response.
       * @see {@link https://github.com/graphql/graphql-wg/blob/main/rfcs/DeferStream.md#payload-format} for the DeferStream spec
       */
      hasNext?: boolean;
    };
// ...
export type OperationType = 'subscription' | 'query' | 'mutation' | 'teardown';
// ...
export type AnyVariables = { [prop: string]: any } | void | undefined;
// ...
export interface GraphQLRequest<
  Data = any,
  Variables extends AnyVariables = AnyVariables
> {
  /** Unique identifier for the `GraphQLRequest`.
   *
   * @remarks
   * This is a key that combines the unique key of the `query` and the `variables` into a unique
   * `key` for the `GraphQLRequest`. Any request with the same query and variables will have a unique
   * `key` by which results and requests can be identified as identical.
   *
   * Internally, a stable, cached `key` is generated for the `DocumentNode` and for the `variables` and
   * both will be combined into a combined `key` which is set here, based on a DJB2 hash,
   *
   * The `variables` will change the key even if they contain a non-JSON reference. If you pass a custom
   * class instance to `variables` that doesn't contain a `toString` or `toJSON` method, a stable but random
   * identifier will replace this class to generate a key.
   */
  key: number;
  /** A GraphQL document to execute against a cache or API.
   *
   * @remarks
   * A `GraphQLRequest` is executed against an operation in a GraphQL document.
   * In `urql`, we expect a document to only contain a single operation that is executed rather than
   * multiple ones by convention.
   */
  query: DocumentNode | TypedDocumentNode<Data, Variables>;
  /** Variables used to execute the `query` document.
   *
   * @remarks
   * The `variables`, based either on the {@link AnyVariables} type or the {@link TypedDocumentNode}'s provided
   * generic, are sent to the GraphQL API to execute a request.
   */
  variables: Variables;
}
// ...
export interface Operation<
  Data = any,
  Variables extends AnyVariables = AnyVariables
> extends GraphQLRequest<Data, Variables> {
  /** The `OperationType` describing the kind of `Operation`.
   *
   * @remarks
   * This is used to describe what to do with the {@link GraphQLRequest} of an {@link Operation} and is set
   * to a GraphQL operation type (`query`, `mutation`, or `subscription`) to start an `Operation`; and to
   * `teardown` to cancel an operation, which either terminates it early or lets exchanges know that no
   * consumer is interested in this operation any longer.
   */
  readonly kind: OperationType;
  /** Holds additional metadata for an `Operation` used to execute it.
   *
   * @remarks
   * The {@link OperationContext} is created by the {@link Client} but may also be modified by
   * {@link Exchange | Exchanges} and is used as metadata by them.
   */
  context: OperationContext;
}
// ...
export interface OperationResult<
  Data = any,
  Variables extends AnyVariables = AnyVariables
> {
  /** The [operation]{@link Operation} which has been executed. */
  /** The `Operation` which this `OperationResult` is for.
   *
   * @remarks
   * The `operation` property is set to the {@link Operation} that this result is. At the time the
   * {@link OperationResult} is constructed (either from the cache or an API response) the original
   * `Operation` that the exchange delivering this result has received will be added.
   *
   * The {@link Client} uses this to identify which {@link Operation} this {@link OperationResult} is
   * for and to filter and deliver this result to the right place and consumers.
   */
  operation: Operation<Data, Variables>;
  /** The result of the execution of the GraphQL operation.
   * @see {@link https://spec.graphql.org/October2021/#sec-Data} for the GraphQL Data Response spec
   */
  data?: Data;
  /** Contains a description of errors raised by GraphQL fields or the request itself by the API.
   *
   * @remarks
   * The `error` of an `OperationResult` is set to a {@link CombinedError} if the GraphQL API response
   * contained any GraphQL errors.
   *
   * GraphQL errors occur when either a GraphQL request was prevented from executing entirely
   * (at which point `data: undefined` is set) or when one or more fields of a GraphQL request
   * failed to execute. Due to the latter, you may receive partial data when a GraphQL request
   * partially failed.
   */
  error?: CombinedError;
  /** Additional metadata that a GraphQL API may choose to send that is out of spec.
   * @see {@link https://spec.graphql.org/October2021/#sel-EAPHJCAACCoGu9J} for the GraphQL Response spec
   */
  extensions?: Record<string, any>;
  /** Indicates that an `OperationResult` is not fresh and a new result will follow.
   *
   * @remarks
   * The `stale` flag indicates whether a result is expected to be superseded by a new result soon.
   * This flag is set whenever a new result is being awaited and will be deliverd as soon as the API responds.
   *
   * It may be set by the {@link Client} when the `Operation` was already active, at which point
   * the {@link Client} asks the {@link Exchange | Exchanges} to request a new API response, or
   * by cache exchanges when a temporary, incomplete, or initial cache result has been deliverd, and
   * a new API request has been started in the background. (For partial cache results)
   *
   * Most commonly, this flag is set for a cached result when the operation is executed using the
   * `cache-and-network` {@link RequestPolicy}.
   */
  stale?: boolean;
  /** Indicates that the GraphQL response is streamed and updated results will follow.
   *
   * @remarks
   * Due to incremental delivery, an API may deliver multiple {@link ExecutionResult | ExecutionResults} for a
   * single GraphQL request. This can happen for `@defer`, `@stream`, or `@live` queries, which allow an API
   * to update an initial GraphQL response over time, like a subscription.
   *
   * For GraphQL subscriptions, this flag will always be set to `true`.
   */
  hasNext?: boolean;
}